refactor: #5165 move Lru out of PhSticky into its own file
